refactor: using uint8_t (1 byte) instead of int (8 bytes)

+++ c/bitwise.c
@@ -1,8 +1,9 @@
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 
-void primitiveFunctions(int a, int b) {
+void primitiveFunctions(uint8_t a, uint8_t b) {
   printf("primitive functions between %08b and %08b:\n", a, b);
   printf("AND: %d (%08b)\n", a & b, a & b);
   printf("OR: %d (%08b)\n", a | b, a | b);
@@ -11,24 +12,24 @@ void primitiveFunctions(int a, int b) {
   printf("RIGHT SHIFT 2: %d (%08b)\n", a >> 2, a >> 2);
 }
 
-int flipBit(int bits, int position) {
-  // @TODO: flipBit(int bits, int position)
+uint8_t flipBit(uint8_t bits, uint8_t position) {
+  // @TODO: flipBit(uint8_t bits, uint8_t position)
   return bits;
 }
 
-bool isEven(int bits) {
-  // @TODO: isEven(int bits)
+bool isEven(uint8_t bits) {
+  // @TODO: isEven(uint8_t bits)
   return bits;
 }
 
-bool isPowerOfTwo(int bits) {
-  // @TODO: isPowerOfTwo(int bits)
+bool isPowerOfTwo(uint8_t bits) {
+  // @TODO: isPowerOfTwo(uint8_t bits)
   return bits;
 }
 
 int main(int argc, char *argv[]) {
-  int a = 0b00101010; // 42 in binary
-  int b = 0b01011011; // 91 in binary
+  uint8_t a = 0b00101010; // 42 in binary
+  uint8_t b = 0b01011011; // 91 in binary
   primitiveFunctions(a, b);
   return EXIT_SUCCESS;
 }
